split conversion files and add map type parsing
